consolidated test helper code
