+++ assets/__tests__/view_test.js
@@ -1,36 +1,5 @@
 import LiveSocket, { View } from '../js/phoenix_live_view';
-import waitForExpect from 'wait-for-expect';
-import jsdom from 'jsdom';
-const { JSDOM } = jsdom;
-
-function wait(callback = () => {}, {timeout = 4500, interval = 50} = {}) {
-  return waitForExpect(callback, timeout, interval)
-}
-
-function liveViewDOM() {
-  const div = document.createElement('div')
-  div.setAttribute('data-phx-view', '');
-  div.setAttribute('data-phx-session', 'abc123');
-  div.setAttribute('id', 'container');
-  div.innerHTML = `
-    <label for="plus">Plus</label>
-    <input id="plus" value="1" />
-    <button phx-click="inc_temperature">Inc Temperature</button>
-  `;
-  const button = div.querySelector('button')
-  const input = div.querySelector('input')
-  button.addEventListener('click', () => {
-    setTimeout(() => {
-      input.value += 1;
-    }, 200);
-  });
-
-  return div;
-}
-
-function dom() {
-  return new JSDOM(`<!DOCTYPE html><body>${liveViewDOM().outerHTML}</body>`);
-}
+import { liveViewDOM, dom } from './helpers'
 
 test('sets defaults', async () => {
   let liveSocket = new LiveSocket('/live');
